Removed combo system from game (unused) Added dancing blades multiplier

+++ Assets/Scripts/Battle System/Base/Spells/SpellDefinition.cs
@@ -23,7 +23,7 @@ namespace BattleSystem
         
         [Header("Damage")]
         public int Damage;
-        public int DamageMultiplier;
+        public float DamageMultiplier;
         public float DamageDisplacement;
         public DamageType DamageType = DamageType.Spell;
